\name{NanoStringData}
\alias{NanoStringData}
\docType{data}
\title{
A real 'NanoStringSet' object.
}
\description{
The object is created based on Mori Data with normal and tumor groups and 2
samples in each group. The object contain 599 endogenes, 6 pisitive control,
6 negative control and 4 housekeeping control.
}
\usage{data(NanoStringData)}
\value{An object of NanoStringSet}
\examples{
data(NanoStringData)
NanoStringData
}
\keyword{datasets}
