\name{NanoStringDiff-package}
\docType{package}
\alias{NanoStringDiff}
\alias{NanoStringDiff-package}

\title{
 NanoStringDiff package for differential expression analysis 
 of NanoString nCounter data
}
\description{
  NanoStringDiff is an R package for differential expression analysis of 
  NanoString nCounter data, and the main function for differential analysis 
  is \code{\link{glm.LRT}}. See the examples at \code{\link{glm.LRT}} for 
  basic analysis steps. 
  NanoStringDiff utilizes a generalized linear model (GLM) of the negative 
  binomial family to characterize count data and allows for multi-factor design.
}

\author{
  hong Wang <hong.wang@uky.edu>
  chi wang <chi.wang@uky.edu>
}

\keyword{package}
